casing switch value for openmpi

+++ src/clib/pio_darray.c
@@ -475,7 +475,10 @@ int PIOc_write_darray(int ncid, int varid, int ioid, PIO_Offset arraylen, void *
 #endif /* _NETCDF4 */
             vtype = (MPI_Datatype)iodesc->basetype;
             LOG((3, "caller did not provide fill value vtype = %d", vtype));
-            switch (vtype)
+
+            /* Need to cast MPI_Datatype to int in switches or openmpi
+             * fails. */
+            switch ((int)vtype)
             {
             case MPI_BYTE:
                 fill = &byte_fill;
